+++ lib/doppler_nanten_old.py
@@ -2,11 +2,9 @@
 #-*- coding: utf-8 -*-
 
 #import csv
-#from pyslalib import slalib
+from pyslalib import slalib
 from astropy.coordinates import SkyCoord
 from astropy.coordinates import FK5
-from astropy.coordinates import EarthLocation
-from astropy.time import Time
 import astropy.units as u
 import astropy._erfa as erfa
 import math
@@ -94,9 +92,6 @@ class doppler_nanten (object):
                   #"APPARENT"  : 10,
                   #"HORIZONTAL": 100,
                   "same"      : 0}
-    
-    #time
-    t = Time.now()
 
 
 
@@ -112,6 +107,7 @@ class doppler_nanten (object):
             r_dict[record[0]] =record[1]
             continue
         """
+
         pass
 
     #def set_track(self, x, y, coord, offset_x, offset_y, offset_dcos, offset_coord, stime):
@@ -119,9 +115,9 @@ class doppler_nanten (object):
         """
         setting 2ndLO
         """
-        t_now = Time.now().unix
-        self.t = Time(t_now + stime, format = "unix")
-        vobs = self.get_vobs(math.radians(x), math.radians(y), coord,
+        mjd = 40587.0 + time.time()/(24.*3600.)
+        vobs_mjd = mjd + stime/24./3600.
+        vobs = self.get_vobs(vobs_mjd,math.radians(x),math.radians(y),coord,
                              offset_x, offset_y, offset_dcos, offset_coord)
         c = self.dic1["LIGHT_SPEED"]
         for band in range(1, self.dic1["bandnum"]+1):
@@ -172,7 +168,7 @@ class doppler_nanten (object):
         setting 2ndLO
         """
         mjd = mjd + secofday / (24.* 3600.)
-        vobs = self.get_vobs(math.radians(x),math.radians(y),coord,
+        vobs = self.get_vobs(vobs_mjd,math.radians(x),math.radians(y),coord,
                              offset_x, offset_y, offset_dcos, offset_coord)
         c = self.dic1["LIGHT_SPEED"]
         for band in range(1, self.dic1["bandnum"]+1):
@@ -216,8 +212,8 @@ class doppler_nanten (object):
         power = {"sg":pow_sg,}
         return {"freq":freq, "power":power}
     """
-    def get_vobs(self, xtmp, ytmp, mode, offx, offy, dcos, offmode):
-        #mode convert to j2000
+    def get_vobs(self, mjdtmp, xtmp, ytmp, mode, offx, offy, dcos, offmode):
+        #modeを見て、j2000へ座標変換してる
         mode = mode.lower()
         offmode = offmode.lower()
         mode = self.coord_dict[mode]
@@ -241,15 +237,11 @@ class doppler_nanten (object):
             xxtmp = xtmp
             yytmp = ytmp
         elif mode == self.coord_dict["b1950"] :
-            sc_fk4 = SkyCoord( ra = xtmp*u.radian, dec = ytmp*u.radian, frame = "fk4")
-            sc_fk5 = sc_fk4.fk5
-            ret = [sc_fk5.ra.radian, sc_fk5.dec.radian]
+            ret = slalib.sla_fk45z(xtmp, ytmp, 1950)
             xxtmp = ret[0]
             yytmp = ret[1]
         elif mode == self.coord_dict["lb"] or mode == self.coord_dict["galactic"] or mode == self.coord_dict["gal"]:
-            sc_gal = SkyCoord( ra = xtmp*u.radian, dec = ytmp*u.radian, frame = "galactic")
-            sc_fk5 = sc_gal.fk5
-            ret = [sc_fk5.ra.radian, sc_fk5.dec.radian]
+            ret = slalib.sla_galeq(xtmp, ytmp)
             xxtmp = ret[0]
             yytmp = ret[1]
         else :
@@ -258,85 +250,199 @@ class doppler_nanten (object):
         print(xxtmp, yytmp)
         print(xxtmp/3600.)
         print(xxtmp*360./2*3.14)
-        vobs = self.calc_vobs(xxtmp, yytmp)
+        vobs = self.calc_vobs(mjdtmp+2400000.5, xxtmp, yytmp)
         print('vobs',vobs,type(vobs))
         return vobs
 
-    def calc_vobs(self, ra_2000, dec_2000):
-        
-        x_2000 = []
-        x = []
-        x_solar = []
-        v_earth = []
-        v_solar = []
-        solar_tmp = []
+    def calc_vobs(self, jd, ra_2000, dec_2000):
 
+        x_2000 = [0.,0.,0.]
+        x = [0.,0.,0.]
+        x1 = [0.,0.,0.]
+        v = [0.,0.,0.]
+        v_rev = [0.,0.,0.]
+        v_rot = [0.,0.,0.]
+        v2 = [0.,0.,0.]
+        solx = [0.,0.,0.]
+        solx1 = [0.,0.,0.]
+        solv = [0.,0.,0.]
+        DEG2RAD = math.pi/180.
+        SEC2RAD = (2*math.pi)/(24.*60.*60.)
+        ARCSEC2RAD = math.pi/(180.*60.*60.)
+        RAD2DEG = 180./math.pi
         #1.85m at nobeyama
-        #glongitude = 138.472153
+        #glongitude = 138.472153 * math.pi/180.
         #nanten2 at atacama (nanten2wiki)
-        glongitude = -67.70308139
+        glongitude = -67.70308139 * DEG2RAD
         #1.85m at nobeyama
-        #glatitude = 35.940874
+        #glatitude = 35.940874 * math.pi/180.
         #nanten2 at atacama (nanten2wiki)
-        glatitude = -22.96995611 
+        glatitude = -22.96995611 * DEG2RAD
         #1.85m at nobeyama
         #gheight = 1386
         #nanten2 at atacama
         gheight = 4863.85
 
-        #init
+        #gdut1 = -0.14
+        #gstop_flag = 1
+        #gkisa = [0,0,0,0,0,0,0,0,0,0,0,0,0]
+        #gtemperature = 273.0
+        #gpressure = 860
+        #ghumidity = 0.2
+        #gtlr = 0.0065
+        #tai_utc = 34.0
+        #instruction = 0
+        #target_changed = 1
+
+
+
+
+        #ra_2000=ra_2000*DEG2RAD
+        #dec_2000=dec_2000*DEG2RAD
         a = math.cos(dec_2000)
-        x_2000 = [a * math.cos(ra_2000), a * math.sin(ra_2000), math.sin(dec_2000)]
+        x_2000[0] = a*math.cos(ra_2000)
+        x_2000[1] = a*math.sin(ra_2000)
+        x_2000[2] = math.sin(dec_2000)
 
-        tu = (self.t.jd - 2451545.) / 36525.
-        #correction(precession, nutation)
-        sc_fk5 = SkyCoord(ra = ra_2000*u.radian, dec = dec_2000*u.radian, frame = "fk5")
-        trans = sc_fk5.transform_to(FK5(equinox = "J{}".format(2000.+tu*100.)))
-        ranow = trans.ra.radian
-        delnow = trans.dec.radian
+        tu= (jd - 2451545.) / 36525.
         
-        x = self.nutation(ranow, delnow)
-
-        #Earth velocity to object
-        earthloc = EarthLocation(lon = glongitude*u.deg, lat = glatitude*u.deg, height = gheight*u.m)
-        earthloc_gcrs = earthloc.get_gcrs(self.t)
-        sc_earth = SkyCoord(earthloc_gcrs).fk5
-
-        v_earth = sc_earth.velocity.d_xyz.value
-
-        vobs = - (v_earth[0] * x_2000[0] + v_earth[1] * x_2000[1] + v_earth[2] * x_2000[2])
-
-        #solar
-        sc_fk4_sol = SkyCoord(ra = 18*15*u.deg, dec = 30*u.deg, frame = "fk4")
-        trans_sol = sc_fk4_sol.transform_to(FK5(equinox = "J{}".format(2000.+tu*100.)))
-        rasol = trans_sol.ra.radian
-        delsol = trans_sol.dec.radian
+        ranow=ra_2000
+        delnow=dec_2000
+        ret =slalib.sla_preces("FK5", 2000.,2000.+tu*100.,ranow,delnow)
+        ranow = ret[0]
+        delnow = ret[1]
         
-        solar_tmp = self.nutation(rasol, delsol)
-        v_solar = [solar_tmp[0]*20, solar_tmp[1]*20, solar_tmp[2]*20]
+        a = math.cos(delnow)
+        x[0]=a*math.cos(ranow)
+        x[1]=a*math.sin(ranow)
+        x[2]=math.sin(delnow)
         
-        vobs = -(vobs - (v_solar[0] * x[0] + v_solar[1] * x[1] + v_solar[2] * x[2]))
-
-
-        return vobs
-
-    def nutation(self, ra, dec):
-        
-        pos = []
-
-        a = math.cos(dec)
-        tmp = [a * math.cos(ra), a * math.sin(ra), math.sin(dec)]
-        
-        #correction(nutation)
-        ret = erfa.nut00a(self.t.jd, 0.0)
+        ret = slalib.sla_nutc(jd-2400000.5)#radian
         nut_long = ret[0]
         nut_obliq = ret[1]
+        eps0 = ret[2]
         
-        pos.append(tmp[0] - (tmp[1] * math.cos(nut_obliq) + tmp[2] * math.sin(nut_obliq)) * nut_long)
-        pos.append(tmp[1] + tmp[0] * math.cos(nut_obliq) * nut_long - tmp[2] * nut_obliq)
-        pos.append(tmp[2] + tmp[0] * math.sin(nut_obliq) * nut_long + tmp[1] * nut_obliq)
+        #x1[0]=x[0]-(x[1]*math.cos(eps0)+x[2]*math.sin(eps0))*nut_long
+        #x1[1]=x[1]+x[0]*math.cos(eps0)*nut_long - x[2]*nut_obliq
+        #x1[2]=x[2]+x[0]*math.sin(eps0)*nut_long + x[1]*nut_obliq
 
-        return pos
+
+        x1[0]=x[0]-(x[1]*math.cos(nut_obliq)+x[2]*math.sin(nut_obliq))*nut_long
+        x1[1]=x[1]+x[0]*math.cos(nut_obliq)*nut_long - x[2]*nut_obliq
+        x1[2]=x[2]+x[0]*math.sin(nut_obliq)*nut_long + x[1]*nut_obliq
+
+        x[0]=x1[0] #x
+        x[1]=x1[1] #y
+        x[2]=x1[2] #z
+
+
+        v0= 47.404704e-3
+
+        ramda=35999.3729*tu+100.4664+(1.9146-0.0048*tu)*math.cos((35999.05*tu+267.53)*DEG2RAD)+0.0200*math.cos((71998.1*tu+265.1)*DEG2RAD)
+
+        r=1.000141+(0.016707-0.000042*tu)*math.cos((35999.05*tu+177.53)*DEG2RAD)+0.000140*math.cos((71998.*tu+175.)*DEG2RAD)
+
+        ramda1=628.308+(20.995-0.053*tu)*math.cos((35999.5*tu+357.52)*DEG2RAD)+0.439*math.cos((71998.1*tu+355.1)*DEG2RAD)+0.243*math.cos((445267.*tu+298.)*DEG2RAD)
+
+        beta=0.024*math.cos((483202.*tu+273.)*DEG2RAD)
+
+        r1 = (10.497-0.026*tu) * math.cos((35999.05*tu+267.53)*DEG2RAD)+ 0.243 * math.cos((445267.*tu+28.)*DEG2RAD)+ 0.176 * math.cos((71998.*tu+265.)*DEG2RAD)
+
+        ramda = ramda *DEG2RAD
+
+        v[0] = -r*ramda1*math.sin(ramda) + r1*math.cos(ramda)
+        v[1] = r*ramda1*math.cos(ramda) + r1*math.sin(ramda)
+        v[2] = r * beta
+
+        v[0] = v[0] - (0.263 * math.cos((3034.9*tu+124.4)*DEG2RAD) +0.058 * math.cos((1222. *tu+140.)*DEG2RAD) +0.013 * math.cos((6069. *tu+144.)*DEG2RAD))
+        v[1] = v[1] - (0.263 * math.cos((3034.9*tu+34.4)*DEG2RAD) +0.058 * math.cos((1222. *tu+50.)*DEG2RAD) +0.013 * math.cos((6069. *tu+54.)*DEG2RAD))
+
+        v[0] = v[0] * v0
+        v[1] = v[1] * v0
+        v[2] = v[2] * v0
+
+        e = (23.439291 - 0.013004*tu)*3600.
+        v_rev[0] = v[0]
+        v_rev[1] = v[1] * math.cos(e * ARCSEC2RAD) - v[2] * math.sin(e * ARCSEC2RAD)
+        v_rev[2] = v[1] * math.sin(e * ARCSEC2RAD) + v[2] * math.cos(e * ARCSEC2RAD)
+
+        v_e = (465.1e-3) * (1.+0.0001568*gheight/1000.) * math.cos(glatitude)/math.sqrt(1.+0.0066945*math.pow(math.sin(glatitude),2.0))
+
+        am = 18.*3600.+41.*60.+50.54841+8640184.812866*tu +0.093104*tu*tu-0.0000062*tu*tu*tu
+        gmst = (jd - 0.5 - int(jd - 0.5)) * 24. * 3600. + am - 12.*3600.
+
+        l = 280.4664*3600. + 129602771.36*tu - 1.093*tu*tu
+        l = l * ARCSEC2RAD
+        p = (282.937+1.720*tu)*3600.
+        p = p * ARCSEC2RAD
+
+        w = (125.045 - 1934.136*tu + 0.002*tu*tu)*3600.
+        w = w * ARCSEC2RAD
+        ll = (218.317+481267.881*tu-0.001*tu*tu)*3600.
+        ll = ll*ARCSEC2RAD
+        pp = (83.353+4069.014*tu-0.010*tu*tu)*3600.
+        pp = pp*ARCSEC2RAD
+        dpsi = (-17.1996-0.01742*tu)*math.sin(w) + (-1.3187)*math.sin(2*l)+0.2062*math.sin(2*w) +0.1426*math.sin(l-p)-0.0517*math.sin(3*l-p)+0.0217*math.sin(l+p) +0.0129*math.sin(2*l-w)-0.2274*math.sin(2*ll)+0.0712*math.sin(ll-pp) -0.0386*math.sin(2*ll-w)-0.0301*math.sin(3*ll-pp) -0.0158*math.sin(-ll+3*l-pp)+0.0123*math.sin(ll+pp)
+        e = e  * ARCSEC2RAD
+
+        dpsicose = dpsi * math.cos(e)
+
+        lst = gmst + (dpsicose + glongitude*RAD2DEG*3600.) / 15.
+
+        v_rot[0] = -v_e * math.sin(lst * SEC2RAD)
+        v_rot[1] = v_e * math.cos(lst * SEC2RAD)
+        v_rot[2] = 0.
+
+        v2[0] = v_rev[0] + v_rot[0]
+        v2[1] = v_rev[1] + v_rot[1]
+        v2[2] = v_rev[2] + v_rot[2]
+        vobs = -(v2[0] * x_2000[0] + v2[1] * x_2000[1] + v2[2] * x_2000[2])
+        
+        rasol=18.*15. *DEG2RAD
+        delsol=30.*DEG2RAD
+
+        #ret = slalib.sla_preces( "FK4", 1900.,2000.+tu*100., rasol, delsol)
+
+        ret = slalib.sla_preces( "FK4", 1950.,2000.+tu*100., rasol, delsol)
+
+        rasol = ret[0]
+        delsol = ret[1]
+
+        a = math.cos(delsol)
+        solx[0] = a*math.cos(rasol)
+        solx[1] = a*math.sin(rasol)
+        solx[2] = math.sin(delsol)
+
+
+        #solx1[0] = solx[0] - (solx[1] * math.cos(eps0) + solx[2] *	\
+        #		      math.sin(eps0)) * nut_long
+        #solx1[1] = solx[1] + (solx[0] * math.cos(eps0) * nut_long\
+        #		      - solx[2] * nut_obliq)
+        #solx1[2] = solx[2] + (solx[0] * math.sin(eps0) * nut_long \
+        #		      + solx[1] * nut_obliq)
+
+
+        solx1[0] = solx[0] - (solx[1] * math.cos(nut_obliq) + solx[2] * math.sin(nut_obliq)) * nut_long
+        solx1[1] = solx[1] + (solx[0] * math.cos(nut_obliq) * nut_long - solx[2] * nut_obliq)
+        solx1[2] = solx[2] + (solx[0] * math.sin(nut_obliq) * nut_long + solx[1] * nut_obliq)
+
+        solv[0]=solx1[0]*20.
+        solv[1]=solx1[1]*20.
+        solv[2]=solx1[2]*20.
+        vobs = vobs - (solv[0] * x[0] + solv[1] * x[1] + solv[2] * x[2])
+        vobs = -vobs
+
+        #print("vobs=%f\n" % vobs)
+        """
+        if gcalc_flag == 1 :
+        return vobs
+        else gcalc_flag == 2:
+        return lst
+        """
+        
+	#print('vobs',vobs,type(vobs)) km/s
+        return vobs
+    
 
 if __name__=="__main__":
     dp = doppler_nanten()
